Hovering selectable area
The previous code had a small issue where a portion of the menu was not hoverable while moving the cursor from the parent menu to the submenu. This caused the submenu to start closing when the cursor left the parent selection for the dropdown submenu section. To fix this, I added positive padding and a negative margin to create a selectable area between the parent and submenu. 

- Added padding/margin to create a selectable area
- Moving the submenu up (from 150% to 100%)
- Submenu li decreased the margins